move msg server to base swap module; msg server doesn't need private access to keeper module -- should use exported functionality only
